+++ e2e/interactive/init.interactive.e2e.3.js
@@ -17,9 +17,6 @@ chai.use(require('chai-fs'));
 describe('run bit init - interactive', function () {
   this.timeout(0);
   const helper = new Helper();
-  after(() => {
-    helper.destroyEnv();
-  });
   describe('with defaults', () => {
     // Skip on windows since the interactive keys are not working on windows
     if (IS_WINDOWS || process.env.APPVEYOR === 'True') {
@@ -139,24 +136,26 @@ describe('run bit init - interactive', function () {
       });
       after(() => {
         helper.restoreConfigs(configsBackup);
+        helper.destroyEnv();
       });
       it('should prefer interactive.init config over interactive config', async () => {
         helper.setConfig(CFG_INTERACTIVE, true);
         helper.setConfig(CFG_INIT_INTERACTIVE, false);
-        const output = await helper.initInteractive([]);
+        const output = helper.initWorkspace();
         // We didn't enter anything to the interactive but we don't expect to have it so the workspace should be initialized
         expect(output).to.have.string('successfully initialized');
       });
-      it('should should not show interactive when interactive config set to false', async () => {
-        helper.setConfig(CFG_INTERACTIVE, false);
-        const output = await helper.initInteractive([]);
-        // We didn't enter anything to the interactive but we don't expect to have it so the workspace should be initialized
-        expect(output).to.have.string('successfully initialized');
-      });
-      it('should should show interactive by default', async () => {
-        const output = await helper.initInteractive(inputsWithDefaultsNoCompiler);
+      it('should should show interactive when interactive config set to true', async () => {
+        helper.setConfig(CFG_INTERACTIVE, true);
+        const output = helper.initWorkspace();
         // We don't enter anything we just want to see that any question has been asked
         expect(output).to.have.string(PACKAGE_MANAGER_MSG_Q);
+      });
+      it('should not show interactive by default', async () => {
+        helper.delConfig(CFG_INTERACTIVE);
+        helper.delConfig(CFG_INIT_INTERACTIVE);
+        const output = helper.initWorkspace();
+        expect(output).to.have.string('successfully initialized');
       });
     }
   });